vulkan tutorial engine as base to make certain things easier (basically having a good setup to build from). Removed code that is replaced by the vulkan tut code so there's no overlap.
also added SDL, and some other third party stuff. Removed Boost (since it was a pain in the ass anyway, and doesn't provide enough value for me at the moment). Did some more coding on the serialization system to introduce a "assetlist" type, which basically loads other assets and asset lists. Improved the class system a bit as well. Added a globals file for global pointers that should be accessible anywhere for ease of use.